yolov5/conda-env-testing/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
ultralytics>=8.2.34
pillow>=10.3.0
numpy>=1.23.5
gitpython>=3.1.30
pandas>=1.1.4
